7/2 회의내용 반영
Fix: 오타 수정
Add: ResultResponseDto 의 타입 변경
Add: Projcet Class에 빌더 패턴 추가

+++ src/main/java/io/seoul/helper/controller/dto/ResultResponseDto.java
@@ -7,12 +7,12 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 public class ResultResponseDto<T> {
-    private Long statusCode;
-    private Long count;
+    private Integer statusCode;
+    private Integer count;
     private T data;
 
     @Builder
-    public ResultResponseDto(Long statusCode, Long count, T data) {
+    public ResultResponseDto(Integer statusCode, Integer count, T data) {
         this.statusCode = statusCode;
         this.count = count;
         this.data = data;
